refactor: xmodule/xml_module.py -> xmodule/xml_block.py
